+++ core/src/lib.rs
@@ -17,6 +17,7 @@ pub mod codegen;
 pub mod error;
 mod from_field;
 mod from_derive_input;
+mod from_generics;
 mod from_meta_item;
 mod from_type_param;
 mod from_variant;
@@ -26,11 +27,12 @@ pub mod util;
 pub use error::{Result, Error};
 pub use from_derive_input::FromDeriveInput;
 pub use from_field::FromField;
+pub use from_generics::FromGenerics;
 pub use from_meta_item::{FromMetaItem};
 pub use from_type_param::FromTypeParam;
 pub use from_variant::FromVariant;
 
 #[cfg(test)]
 mod tests {
-    
+
 }